Unit1_FinalProject.ipynb
The goal of our project was to learn more about the seismic activity in the South American region. 
First we mounted our drives, imported matplotlib.pyplot as plt and read in the text files named all_quakes. To separate the items in each line of the data we used .split(). Longitude and latitude are needed to create the scatter plot so we created empty lists. To loop through the lines in quakes_list and append the latitude and longitude if the earthquake for that row occurred in our region and during our year range. Our years and regions were from 1970 to 2007 in the South American Region. 
After creating the scatterplots we concluded that our region has the most frequent earthquakes by the plates.